fix: SpaceGrotesk font registration
